+++ RELEASE_NOTES.md
@@ -28,3 +28,4 @@
 * Fixed an issue where the Voice Fight input device was reset after restarting.
 * Added missing i18n.
 * Fixed an issue when spinning while using a FrameShot frame it was causing to flip around.
+* Fixed an issue when using GIF record in FrameShot for longer than 8 seconds resulting in another recording.